Added BaseViewModel and refactor packageing
