--- app/src/main/java/com/remotearth/fake_coder/chatapp/SignUpViewModel.kt
@@ -1,7 +1,0 @@
-package com.remotearth.fake_coder.chatapp
-
-import androidx.lifecycle.ViewModel
-
-class SignUpViewModel : ViewModel() {
-
-}
