+++ app/src/main/java/com/remotearth/fake_coder/chatapp/contracts/BaseView.kt
@@ -1,4 +1,4 @@
-package com.remotearth.fake_coder.chatapp
+package com.remotearth.fake_coder.chatapp.contracts
 
 interface BaseView {
     fun isInternetAvailable(): Boolean?
