+++ app/src/main/java/com/remotearth/fake_coder/chatapp/screens/activity/MainActivity.kt
@@ -1,7 +1,8 @@
-package com.remotearth.fake_coder.chatapp
+package com.remotearth.fake_coder.chatapp.screens.activity
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.remotearth.fake_coder.chatapp.R
 
 class MainActivity : AppCompatActivity() {
