+++ app/src/main/java/com/remotearth/fake_coder/chatapp/viewModels/LoginViewModel.kt
@@ -1,0 +1,7 @@
+package com.remotearth.fake_coder.chatapp.viewModels
+
+import com.remotearth.fake_coder.chatapp.viewModels.base.BaseViewModel
+
+class LoginViewModel : BaseViewModel() {
+
+}
