+++ app/src/main/java/com/remotearth/fake_coder/chatapp/viewModels/SignUpViewModel.kt
@@ -1,0 +1,7 @@
+package com.remotearth.fake_coder.chatapp.viewModels
+
+import com.remotearth.fake_coder.chatapp.viewModels.base.BaseViewModel
+
+class SignUpViewModel : BaseViewModel() {
+
+}
