+++ app/src/main/java/com/remotearth/fake_coder/chatapp/viewModels/base/BaseViewModel.kt
@@ -1,0 +1,17 @@
+package com.remotearth.fake_coder.chatapp.viewModels.base
+
+import androidx.databinding.ObservableField
+import androidx.lifecycle.ViewModel
+
+open class BaseViewModel: ViewModel() {
+
+    private var isLoading = ObservableField<Boolean>()
+
+    fun showLoader() {
+        isLoading.set(true)
+    }
+
+    fun hideLoader() {
+        isLoading.set(false)
+    }
+}
