+++ app/src/main/res/layout/activity_main.xml
@@ -8,7 +8,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
 
-    tools:context=".MainActivity">
+    tools:context=".screens.activity.MainActivity">
 
     <fragment
         android:id="@+id/nav_host_fragment"
