+++ app/src/main/res/layout/login_fragment.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".LoginFragment">
+    tools:context=".screens.fragments.LoginFragment">
 
     <TextView
         android:layout_width="match_parent"
