+++ app/src/main/res/layout/sign_up_fragment.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SignUpFragment">
+    tools:context=".screens.fragments.SignUpFragment">
 
     <TextView
         android:layout_width="match_parent"
