+++ app/src/main/res/navigation/navigation_graph.xml
@@ -9,13 +9,13 @@
 
     <fragment
         android:id="@+id/loginFragment"
-        android:name="com.remotearth.fake_coder.chatapp.LoginFragment"
+        android:name="com.remotearth.fake_coder.chatapp.screens.fragments.LoginFragment"
         android:label="login_fragment"
         tools:layout="@layout/login_fragment"/>
 
     <fragment
         android:id="@+id/signUpFragment"
-        android:name="com.remotearth.fake_coder.chatapp.SignUpFragment"
+        android:name="com.remotearth.fake_coder.chatapp.screens.fragments.SignUpFragment"
         android:label="sign_up_fragment"
         tools:layout="@layout/sign_up_fragment"/>
